# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Zoo Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/animals:
    post:
      summary: Create a new animal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalRequest'
      responses:
        '200':
          description: Animal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
    get:
      summary: Get all animals
      responses:
        '200':
          description: List of animals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Animal'
  /api/animals/{animalId}/move:
    post:
      summary: Move an animal to a new enclosure
      parameters:
        - name: animalId
          in: path
          required: true
          schema:
            type: integer
        - name: newEnclosureId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Animal moved successfully
...
components:
  schemas:
...
    Enclosure:
      type: object
      properties:
        id:
          type: integer
        allowedType:
          type: string
        size:
          type: number
        maxCapacity:
          type: integer
        animals:
          type: array
          items:
            $ref: '#/components/schemas/Animal'
    AnimalRequest:
      type: object
      properties:
        name:
          type: string
        species:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
        favoriteFood:
          type: string
        status:
          type: string
      required:
        - name
        - species
        - birthDate
        - gender
        - favoriteFood
        - status
    Animal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        animalSpecies:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
        favoriteFood:
          type: string
        status:
          type: string
        currentEnclosure:
          $ref: '#/components/schemas/Enclosure'
...